+++ packages/antd/src/routing/NavMenuBarAnt.tsx
@@ -11,8 +11,8 @@ import {
     UIFragmentSpec,
     Anchor,
     AnchorProps,
-    useLinkGenerator,
     useLocationManager,
+    useOptionalLinkGenerator,
 } from '@moxb/react-html';
 import { Menu } from 'antd';
 import { observer } from 'mobx-react-lite';
@@ -64,7 +64,7 @@ export interface NavMenuProps<DataType>
  * This widget show an Ant menu bar, based on the state-space.
  */
 export const NavMenuBarAnt = observer((props: NavMenuProps<any>) => {
-    const linkGenerator = useLinkGenerator();
+    const linkGenerator = useOptionalLinkGenerator();
     const locationManager = useLocationManager('nav menu bar ' + props.id);
 
     function getLocationDependantStateSpaceHandler(): LocationDependentStateSpaceHandler<